fix(backbone): Mark the inner entry has empty instead of setting it to null
Setting it to null removes the reference to the inner MutableStateFlow. On the next 'get', a new one is instantiated.

This means that all previous subscribers are now subscribed to a dead StateFlow, and will never receive updates again.

Instead, this commit marks the data as empty instead (which forces the 'get' subscribers to start a new query). However, this means the cache memory is never cleaned (unbounded growth of the cache): to be fixed in a future commit.